Using jQuery.proxy to force "this" to be our object when the callback methods are executed.
This is a really common problem with callback methods, which we don't see at all in PHP. In JavaScript, it's possible that when you say: hey you, execute $foo->doSomething(), that when that method is executed, "$this" is actually something different inside that method. That is unique to JavaScript. In PHP, $this is always $this, but in JavaScript, you can control what "this" is. In jQuery event, this is set to something different (it's set to event.currentTarget). By using jQuery.proxy, we can take control back and force it to be the current object we're in.