Using jQuery.extends to take the exact same changes that we were making to MagicBoxes.prototype and put them into one full object that we then extend all at once.
This is almost like an array merge for objects. I merge on object into another, which adds all the objects from that method into my object. In this case, we're adding a bunch of methods to the "prototype", so that when we instantiate a new instance, that object has all of these methods.